[C++] Fix Arrow stream doc/type inconsistencies
- Correct ArrowStream ref in create_arrow_stream comment
- Document all stream_paused_max_wait_time_ms cases
- Use uint64_t to match core's Option<u64>; cast sentinel explicitly
- Clarify ArrowStreamOptions defaults source

+++ cpp/src/detail/config_convert.hpp
@@ -53,10 +53,13 @@ inline CArrowStreamConfigurationOptions to_c(const ArrowStreamOptions& opts) {
   c.flush_timeout_ms = opts.flush_timeout_ms;
   c.connection_timeout_ms = opts.connection_timeout_ms;
   c.ipc_compression = static_cast<std::int32_t>(opts.ipc_compression);
-  // -1 sentinel = "wait full server duration" (no explicit cap).
+  // The C field is signed with -1 as the "wait full server duration" sentinel;
+  // `nullopt` maps to it. A present value is non-negative (unsigned in the
+  // option), so cast to the signed field explicitly rather than letting the
+  // ternary fold both branches to unsigned.
   c.stream_paused_max_wait_time_ms =
       opts.stream_paused_max_wait_time_ms.has_value()
-          ? *opts.stream_paused_max_wait_time_ms
+          ? static_cast<std::int64_t>(*opts.stream_paused_max_wait_time_ms)
           : -1;
   return c;
 }
